Move Converter to package 'reporting'
